pyproject.toml
[project]
name = "browse-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "playwright>=1.54.0", 
    "python-bidi>=0.6.6",
    "python-dotenv>=1.1.1",
    "httpx>=0.25.0",
]
